Fix python2.7 syntax and old class errors. Async await syntax still needs fixing.
